rpm upgrade failure on EPEL
This `rpm -U` call on the static EPEL rpm will cause system file conflicts,
or installation failures. This depends on the state of whether epel is installed, or not,
on the selected image.

A `yum install -y epel-release` is safer to call in cluster-init.
If you really need to upgrade a baked-in epel to a 6.8 release, or newer. Perhaps
it might be safer to call `yum -y install epel-release || yum -y upgrade epel-release`
in your cluster-init.